fix: Fixed ImHex crashing when using ctrl-backspace on empty file.
Editor was attempting to delete non-existent chars which is UB. Fixed by checking before deleting. Also fixed was a problem created ny having to press enter to change the search string which advanced the selection to the first match. In the next step one would expect that pressing enter on the replace field would replace the selected item but was replacing the item found after he first.

This was fixed by always replacing the current selection first. If the replacement is the same as the searched term then replacing won't advance the cursor, but if they are different then the current match will no longer exist so it would search fora new one.