improv: Added scrolling mouse selections.
The scrolling trigger is limited to the first and last lines of the text editor for now. Also fixed a bug that put the cursor at the beginning of the last line when the text editor was clicked below it. Renamed some variables and fixed some logic problems with moving cursor left and right.